requirements.txt
flask
requests
schedule
python-telegram-bot==13.15
matplotlib